move csv parser into own util file and remove unnecessary console log
